tools/gui_migration.py
"""
GUI Migration Tool for Rush Royale Bot
Helps transition from legacy Tkinter GUI to modern CustomTkinter GUI
"""

import os
import sys
import subprocess
import warnings

# Suppress known harmless warnings
warnings.filterwarnings("ignore", message="pkg_resources is deprecated")

def check_requirements():
    """Check if all requirements are met for modern GUI"""
    print("🔍 Checking requirements for modern GUI...")
    
    # Check Python version
    if sys.version_info < (3, 9):
        print("❌ Python 3.9+ required for CustomTkinter")
        return False
    
    print(f"✅ Python {sys.version_info.major}.{sys.version_info.minor}.{sys.version_info.micro}")
    
    # Check virtual environment
    if not os.path.exists('.venv313'):
        print("❌ Virtual environment '.venv313' not found")
        print("   Run install.bat to create the environment")
        return False
    
    print("✅ Virtual environment found")
    
    # Check if CustomTkinter is available
    try:
        import customtkinter
        print(f"✅ CustomTkinter {customtkinter.__version__} available")
    except ImportError:
        print("⚠️  CustomTkinter not installed - will be installed automatically")
    
    return True

def install_customtkinter():
    """Install CustomTkinter if not already installed"""
    print("\n📦 Installing CustomTkinter...")
    
    # Activate virtual environment and install
    venv_python = os.path.join('.venv313', 'Scripts', 'python.exe')
    
    if os.path.exists(venv_python):
        try:
            result = subprocess.run([venv_python, '-m', 'pip', 'install', 'customtkinter>=5.2.0'], 
                                  capture_output=True, text=True)
            if result.returncode == 0:
                print("✅ CustomTkinter installed successfully")
                return True
            else:
                print(f"❌ Failed to install CustomTkinter: {result.stderr}")
                return False
        except Exception as e:
            print(f"❌ Error installing CustomTkinter: {e}")
            return False
    else:
        print("❌ Virtual environment Python executable not found")
        return False

def compare_guis():
    """Compare features between legacy and modern GUI"""
    print("\n📊 GUI Comparison:")
    print("="*50)
    
    features = [
        ("Appearance", "Basic Tkinter widgets", "Modern dark theme with CTk"),
        ("Layout", "Fixed grid layout", "Responsive frame layout"),
        ("Colors", "Static colors", "Dynamic theming support"),
        ("Buttons", "Standard buttons", "Styled buttons with icons"),
        ("Text Display", "Basic Text widgets", "Modern textboxes with scrolling"),
        ("User Experience", "Functional but dated", "Modern and intuitive"),
        ("Performance", "Good", "Better with optimized rendering"),
        ("Customization", "Limited", "Extensive theming options")
    ]
    
    print(f"{'Feature':<15} {'Legacy GUI':<25} {'Modern GUI':<30}")
    print("-" * 70)
    
    for feature, legacy, modern in features:
        print(f"{feature:<15} {legacy:<25} {modern:<30}")

def test_modern_gui():
    """Test if the modern GUI can be imported and run"""
    print("\n🧪 Testing modern GUI...")
    
    try:
        # Test import
        sys.path.append('Src')
        import gui_modern
        print("✅ Modern GUI imports successfully")
        
        # Test basic functionality
        print("✅ Modern GUI ready to run")
        return True
        
    except ImportError as e:
        print(f"❌ Import error: {e}")
        return False
    except Exception as e:
        print(f"❌ Error testing modern GUI: {e}")
        return False

def migration_summary():
    """Provide migration summary and next steps"""
    print("\n🎯 Migration Summary:")
    print("="*40)
    print("✅ CustomTkinter GUI implementation ready")
    print("✅ Modern styling with dark theme")
    print("✅ Improved layout and user experience")
    print("✅ All original functionality preserved")
    print("✅ Enhanced logging display")
    print("✅ Better error handling")
    
    print("\n🚀 How to use:")
    print("1. Run: launch_modern_gui.bat")
    print("2. Or: python Src\\gui_modern.py")
    print("3. Legacy GUI still available: launch_gui.bat")
    
    print("\n📋 New Features:")
    print("• Dark theme with modern styling")
    print("• Responsive layout that adapts to window size")
    print("• Better organized control panels")
    print("• Enhanced visual feedback")
    print("• Improved error messages")
    print("• Auto-scrolling log display")

def main():
    """Main migration workflow"""
    print("🔄 Rush Royale Bot - GUI Migration Tool")
    print("=" * 50)
    
    # Check requirements
    if not check_requirements():
        print("\n❌ Requirements not met. Please fix the issues above.")
        return False
    
    # Install CustomTkinter if needed
    try:
        import customtkinter
    except ImportError:
        if not install_customtkinter():
            print("\n❌ Failed to install CustomTkinter")
            return False
    
    # Test modern GUI
    if not test_modern_gui():
        print("\n❌ Modern GUI testing failed")
        return False
    
    # Show comparison
    compare_guis()
    
    # Migration summary
    migration_summary()
    
    print("\n✅ Migration completed successfully!")
    print("🎉 Modern GUI is ready to use!")
    
    return True

if __name__ == "__main__":
    try:
        success = main()
        if success:
            print("\n" + "="*50)
            print("To start the modern GUI, run: launch_modern_gui.bat")
            print("="*50)
        else:
            print("\n❌ Migration failed. Check the errors above.")
            
    except KeyboardInterrupt:
        print("\n\n⏹️  Migration cancelled by user")
    except Exception as e:
        print(f"\n❌ Unexpected error during migration: {e}")
    
    input("\nPress Enter to exit...")
